requirements.txt
astropy>=5.3.4,<6
black>=24.8.0,<25
# `jax` and `jaxlib` need to be pinned to the same version
jax==0.4.29
jaxlib==0.4.29
imageio>=2.34.1,<3
# Pinned due to `pipreqs`
ipython==8.12.3
matplotlib>=3.7.0,<4
networkx>=3.3,<4
pillow>=10.3.0,<11
pip-tools>=7.4.1,<8
poliastro>=0.17.0,<1
PuLP>=2.9.0,<3
scipy>=1.13.0,<2
sympy>=1.12.1,<2
